refactor: rename src/const to src/masters
https://github.com/akky-xxxx/eslint-plugin-strict-check/issues/285
